solo falta spllit y bonus

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: alvapere <alvapere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 12:22:32 by alvapere          #+#    #+#             */
-/*   Updated: 2025/04/09 15:22:29 by alvapere         ###   ########.fr       */
+/*   Updated: 2025/04/11 10:20:17 by alvapere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	strlen(const char *s)
 {
-	int		i;
+	size_t	i;
 	char	*ret;
 
 	ret = (char *) s;
@@ -23,3 +23,4 @@ size_t	strlen(const char *s)
 		i++;
 	return (i);
 }
+
